requirements.txt
streamlit
langchain
langchain-community
transformers
sentence-transformers
faiss-cpu
torch
